backend/requirements.txt
Django==1.11.14
django-tastypie==0.14.1
psycopg2==2.7.5
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
six==1.11.0
